Remove unnecessary info from package.json
As this isn't an actual NPM package, it doesn't have a user for most of
the info in package.json. And since the package.json is not the
canonical location for most of this info, it's liable to get out of
date, and shouldn't be trusted. So we should just remove it.